requirements.txt
fastapi
uvicorn
python-multipart
aiohttp
pillow
replicate
openai
python-dotenv
requests
numpy